Make back predictive, and stop the system bars eating the first swipe
The gesture now reports how far through it is and the page moves under the
finger: it draws back from the edge, shrinks a little, takes rounded corners and
a shadow, and the list shows behind it. Let go before the halfway point and it
eases back rather than going anywhere. Android reports the progress itself
through `PredictiveBackHandler` — which needs
`android:enableOnBackInvokedCallback` in the manifest to be sent it at all — and
the iOS edge swipe measures the same thing against the width of the screen.

The list behind a page being dragged away shares the real list's scroll
position, so it is where you left it rather than back at the top.

The first swipe in from an edge did nothing but show the system bars, because
the activity hid them and hidden bars come with
`BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE`. They were only hidden so the pixel
harness would compare content without a status bar over it on one platform and
not the other, so now only the harness hides them — everything else runs with
the bars where the system put them.

Verified on the emulator: the page shrinking and sliding with the list behind on
the first swipe, and all 16 harness comparisons still passing on both platforms.
The iOS gesture is still not script-verifiable — the simulator takes no injected
touches.

+++ src/commonMain/kotlin/com/example/remotecompose/demo/PlatformBack.kt
@@ -3,13 +3,18 @@ package com.example.remotecompose.demo
 import androidx.compose.runtime.Composable
 
 /**
- * Wraps [content] in the platform's own way of going back, for as long as [enabled] holds.
+ * The platform's back gesture, with the progress of one in flight.
  *
- * Android has one and it is the system back gesture. iOS has one for a view controller inside a
- * navigation stack, which a bare Compose view controller is not, so there it is the same
- * left-edge swipe, detected by hand. Desktop has none.
+ * [content] is given 0 when nothing is happening and rises towards 1 as the gesture goes on, so
+ * the screen can move under the finger and show where it is going rather than snapping when the
+ * finger lifts. [onBack] runs when the gesture completes; when it is abandoned the progress falls
+ * back to 0 and nothing else happens.
+ *
+ * Android has a system gesture that reports this, through `PredictiveBackHandler`. iOS has one
+ * for a view controller inside a navigation stack, which a bare Compose view controller is not,
+ * so there it is the same left-edge swipe measured by hand. Desktop has none and stays at 0.
  *
  * Compose Multiplatform grew a common `BackHandler` in 1.8; this project is on 1.7.
  */
 @Composable
-expect fun BackGesture(enabled: Boolean, onBack: () -> Unit, content: @Composable () -> Unit)
+expect fun BackGesture(enabled: Boolean, onBack: () -> Unit, content: @Composable (Float) -> Unit)
